// src/interfaces/returned-data/returned-jumbo-data.ts
export interface ReturnedJumboDataProps {
  id: number;
  attributes: {
    mainContent: string;
    color: string;
    skinCare: string;
    skinCareHeading: string;
    cuts: string;
    stylingAndTreaments: string;
    styleAndTreatmentsHeading: string;
    colorHeading: string;
    cutsHeading: string;
    specialOccasions: string;
    specialOccasionsHeading: string;
    addOns: string;
    addOnsHeading: string;
    nails: string;
    nailsHeading: string;
  };
}